cpp00 remodeled

+++ cpp/cpp00/ex00/megaphone.cpp
@@ -6,7 +6,7 @@
 /*   By: rwintgen <rwintgen@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/08 11:52:58 by rwintgen          #+#    #+#             */
-/*   Updated: 2024/09/16 14:16:14 by rwintgen         ###   ########.fr       */
+/*   Updated: 2024/09/17 11:34:04 by rwintgen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ int main(int argc, char **argv)
 	{
 		std::string tmp = argv[i];
 
-		for (int j = 0; j < (int)tmp.length(); j++)
+		for (size_t j = 0; j < tmp.length(); j++)
 			result += std::toupper(tmp[j]);
 	}
 	std::cout << result << std::endl;
